requirements.txt
pandas>=1.5.0
numpy>=1.21.0

beautifulsoup4>=4.11.0
requests>=2.28.0
selenium>=4.15.0
webdriver-manager>=4.0.0
pyspark>=3.3.0
psycopg2-binary>=2.9.5
dbt-postgres>=1.4.0
piperider>=0.41.0
python-telegram-bot>=20.0
toml>=0.10.2
dash>=2.10.0
dash-bootstrap-components>=1.4.0
plotly>=5.10.0
werkzeug>=2.2.0
flask>=2.2.0
flask-cors>=4.0.0
schedule>=1.2.0
redis>=4.5.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
